Full tabline truncation algorithm
Quite sure it's not the most efficient stuff ever. But the result is nice.
(and the compute time doesn't really explode to be honest. Quite some
looping in the algo, but iterating on small numbers, not on complex
strings)

The algo is an iterative approach. It starts from the active tab that is
always fully displayed, and jumps right and left starting from the
smallest possible tab, expanding them gradually while starting back at
the center regularly ("slope" parameter). If it cannot fit all tabs, a
couple of extra chars are added before/after.

The result is quite smooth. When you move right and left it feels quite
natural, without having weird resizes, and you keep a good overview of
open tabs.